EM-Sense/requirements.txt
Cython
wandb
lightning
torchmetrics
einops
jsonargparse[signatures]